+++ app/src/main/java/de/halfbit/edgetoedge/sample/examples/ConstraintLayoutTransitionsFragment.kt
@@ -8,10 +8,8 @@ import androidx.constraintlayout.widget.ConstraintSet
 import androidx.transition.ChangeBounds
 import androidx.transition.TransitionManager
 import com.bumptech.glide.Glide
-import de.halfbit.edgetoedge.Adjustment
 import de.halfbit.edgetoedge.Edge
 import de.halfbit.edgetoedge.edgeToEdge
-import de.halfbit.edgetoedge.fitEdgeToEdge
 import de.halfbit.edgetoedge.sample.BaseFragment
 import de.halfbit.edgetoedge.sample.R
 import de.halfbit.edgetoedge.sample.commons.QUALITY
@@ -31,19 +29,21 @@ class ConstraintLayoutTransitionsFragment : BaseFragment() {
         Glide.with(requireContext()).load(IMAGE).into(image)
 
         edgeToEdge {
-            tapToExtend.fit {
-                adjustment = Adjustment.Margin
-                Edge.Bottom
-            }
+            tapToExtend.fitMargin { Edge.Bottom }
             description.fit { Edge.Bottom }
         }
 
         tapToExtend.setOnClickListener {
             if (collapse) {
                 transitionInto(R.layout.fragment_constraint_layout_transition)
-                fitEdgeToEdge()
+                edgeToEdge {
+                    tapToExtend.fitMargin { Edge.Bottom }
+                }
             } else {
                 transitionInto(R.layout.fragment_constraint_layout_transition_expanded)
+                edgeToEdge {
+                    tapToExtend.unfit()
+                }
             }
             collapse = collapse.not()
         }