+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 0.6 (unreleased)
 * Fix `consumeInsets` was improperly overwritten in some cases.
 * New `view.fitMargin`, `view.fitPadding` and `view.fitHeight` convenience methods.
+* Internal improvements.
 
 # 0.5 (01.01.2020)
 * New `Fragment.fitEdgeToEdge()` function.